Extracted common methods from concrete controller implementations into AbstractController.php
